// apps/api-core/src/model/enumerations.ts
export enum RepriceResultEnum {
  CHANGE_UP = "CHANGE #UP",
  CHANGE_DOWN = "CHANGE #DOWN",
  CHANGE_ERROR = "CHANGE #ERROR",
  IGNORE_FLOOR = "IGNORE #FLOOR",
  IGNORE_LOWEST = "IGNORE #LOWEST",
  IGNORE_SISTER = "IGNORE #SISTERLOWEST",
  IGNORE_ERROR = "IGNORE #ERROR",
  IGNORE_SETTINGS = "IGNORE #SETTINGS",
  SPECIAL_422 = "422 ERROR",
  DEFAULT = "UNKNOWN",
}